Implement regression prevention for Ts_Segments
Specifically, add support to Ts_RegressionPreventerBatchAccess and
TsRegressionPreventer::_SegmentSolver. Overloads of the methods
IsSegmentRegressive and ProcessSegment now operate on a Ts_Segment.
ProcessSegment modifies the Ts_Segment in place.

In practice, these methods will only be called internally by eval,
sample, bake, and breakdown code to ensure that the segment they are
about to operate on is non-regressive. Currently this code only uses
the "keep ratio" anti-regression mode. In practice, splines are
de-regressed at authoring time so Ts_Segments should only be regressive
if the user has deliberately disabled anti-regression when authoring
and then wants to eval the spline.

Note: this functionality is only exercised by the test, but it is
proposed to be the basis for exec, breakdown, sample, and bake.

(Internal change: 2389416)