Prototype: Enable UsdSkel Native Instancing in Storm Imaging via Vertex Shader Skinning or we now call it deferred skinning
This is part 1 of the change in pxr/imaging/hd.

Introduce HD_ENABLE_DEFERRED_SKINNING env and
HdSkinningSettings::IsSkinningDeferred() API to enable the feature.

Define all the skinning inputs tokens in hd.
When skinning is deferred, we add them to GetBuiltinPrimvars() for
skinnable prims like mesh and basisCurves.

Add T_VEC2IARRAY to hdGenSchema and HdVec2iArrayDataSource to
dataSourceTypeDefs. This is needed for follow up changes.

(Internal change: 2390038)
(Internal change: 2390117)